rabbit: more precise iterconsume timeout
The iterconsume always set the timeout of kombu to 1 second
even the requested timeout more precise or < 1 second.

This change fixes that.

Related bug: #1400268
Related bug: #1399257

(cherry picked from commit 023b7f44e2ccd77a7e9ee9ee78431a4646c88f13)
